Legacy Addons
Moved the font setup to cl_init.
Removed the use of :Center() withing PerformLayouts,
infinite loops yay.
Rebuilt the legacy addon checker,
Will show if client does not have the addons.
Generates a text file in the data folder that when moved to addons and
changed to .sh will install the missing addons.
Addon checks now has a 5 step tutorial for installing.